# Optimize inlay hints