--- a/src/zls.zig
+++ b/src/zls.zig
@@ -1,6 +1,7 @@
 // Used by tests as a package, can be used by tools such as
 // zigbot9001 to take advantage of zls' tools
 
+pub const ast = @import("ast.zig");
 pub const analysis = @import("analysis.zig");
 pub const Header = @import("Header.zig");
 pub const debug = @import("debug.zig");