--- a/tests/tests.zig
+++ b/tests/tests.zig
@@ -1,6 +1,7 @@
 comptime {
     _ = @import("helper.zig");
 
+    _ = @import("utility/ast.zig");
     _ = @import("utility/offsets.zig");
     _ = @import("utility/position_context.zig");
     _ = @import("utility/uri.zig");