… src/inlay_hints.zig