This data has been produced by sending 66 inlay hint requests across `AstGen.zig`.

|                | total time | time per request  | speedup compared to master :rocket: |
|----------------|------------|-------------------|------------------------------|
| debug-master |    41730ms |      632ms        |               -              |
| debug-pr      |      582ms |        9.82ms     |             71.7x            |
| safe-master  |     5221ms |       79ms        |               -              |
| safe-pr       |      105ms |        1.59ms     |             49.7x            |
| small-master |     7581ms |      115ms        |               -              |
| small-pr      |      124ms |        1.88ms     |             61.1x            |
| fast-master  |     3452ms |       52ms        |               -              |
| fast-pr       |       95ms |        1.44ms     |             36.3x            |

I've also added some helper function in `Ast.zig` for iterating over node children and finding nodes that contain a certain source range. This could be useful for implementing ranged semantic tokens.